requirements/common.txt
Django>2.2,<2.3
djangorestframework-simplejwt==4.1.0
djangorestframework==3.9.2
django-cors-headers==2.5.2
psycopg2-binary==2.8.3
djangorestframework-simplejwt==4.1.0
cryptography==2.7
sentry-sdk==0.11.1